src/requirements.txt
google-cloud-vision==2.2.0